[measureobjectskeleton]: (4) Fixed optional return value causing pydantic error, fixed frontend crash caused due to missing branchpoint_image
